added a figures dir, with empty mapping file
the intention is to record mappings between output filenames with the
index the figure will have in the paper. So e. g. output name is
"bar_plot_counts.svg" and paper index name is "Figure4b.svg".
Will think about the best way to do this, just get the structure and
ideas down for now.